# Conflicts:
#	back-end/composer.lock
#	front-end/package-lock.json
#	front-end/src/App.jsx
#	front-end/src/context/task-context.jsx

Took 3 minutes

+++ back-end/app/Http/Controllers/SubTaskController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\MainTask;
 use App\Models\SubTask;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class SubTaskController extends Controller
 {
@@ -13,7 +15,8 @@ class SubTaskController extends Controller
      */
     public function index()
     {
-        return SubTask::all();
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+        return SubTask::where('user_id', $userId)->get();
     }
 
     /**
@@ -21,13 +24,16 @@ class SubTaskController extends Controller
      */
     public function create(request $request)
     {
-        if (!$request->title || !$request->main_task_id || !$request->user_id) {
-            return response(['error' => 'you are stupid'], 404);
-        }
         try {
+            if (!$request->title || !$request->main_task_id) {
+                return response(['error' => 'you are stupid'], 404);
+            }
+
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+
             $subTask = new SubTask([
                 'title' => $request->title,
-                'user_id' => $request->user_id,
+                'user_id' => $userId,
                 'description' => $request->description,
                 'main_task_id' => $request->main_task_id,
             ]);
@@ -45,7 +51,12 @@ class SubTaskController extends Controller
      */
     public function show(string $id)
     {
-        return SubTask::query()->findOrFail($id);
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+        if ($userId === $id) {
+            return SubTask::query()->findOrFail($id);
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 
     /**
@@ -54,35 +65,80 @@ class SubTaskController extends Controller
     public function edit(string $id, request $request)
     {
         try {
-
-            if (!$request->title || !$request->main_task_id || !$request->user_id) {
+            if (!$request->title || !$request->main_task_id) {
                 return response(['error' => 'you are stupid'], 404);
             }
             $subTask = SubTask::query()->findOrFail($id);
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+            if ($subTask->user_id === $userId) {
+                $subTask->update([
+                    'title' => $request->title ?? $subTask->title,
+                    'user_id' => $userId,
+                    'description' => $request->description ?? $subTask->description,
+                    'main_task_id' => $request->main_task_id ?? $subTask->main_task_id,
+                    'group_id' => $request->group_id ?? $subTask->group_id,
 
-            $subTask->update([
-                'title' => $request->title ?? $subTask->title,
-                'user_id' => $request->user_id ?? $subTask->user_id,
-                'description' => $request->description ?? $subTask->description,
-                'main_task_id' => $request->main_task_id ?? $subTask->main_task_id,
-                'group_id' => $request->group_id ?? $subTask->group_id,
-
-            ]);
-            $subTask->save();
-            return $subTask;
+                ]);
+                $subTask->save();
+                return $subTask;
+            } else {
+                return response()->json(['error' => 'you are not authorized'], 403);
+            }
         } catch (ModelNotFoundException $e) {
             return response(['error' => $e], 500);
         }
     }
 
+    public function completed(string $id, request $request)
+    {
+        if (!$request->has('completed')) {
+            return response(['error' => 'you are stupid'], 400);
+        }
+
+        $subTask = SubTask::query()->findOrFail($id);
+        $subTask->completed = $request->completed;
+
+        if ($request->completed) {
+            $ts = time();
+            $curDate = date('Y-m-d H:i:s', $ts);
+            $subTask->completed_at = $curDate;
+        } else {
+            $subTask->completed_at = null;
+        }
+        $subTask->save();
+
+        $this->progressUpdate($subTask->main_task_id, "1");
+
+        return $subTask;
+    }
+
+    public function progressUpdate(string $id, string $userId)
+    {
+        $mainTask = MainTask::query()->findOrFail($id);
+
+        $subtasks = SubTask::query()->where('main_task_id', '=', $id)->count();
+        $completedSubtasks = SubTask::query()->where('main_task_id', '=', $id)->where('completed', '=', 1)->count();
+
+        $progress = $completedSubtasks / $subtasks * 100;
+
+        $mainTask->users()->updateExistingPivot($userId, [
+            'progress' => $progress,
+        ]);
+    }
 
     /**
      * Remove the specified resource from storage.
      */
     public function delete(string $id)
     {
+        $userId = JWTAuth::parseToken()->authenticate()->id;
         $subTask = SubTask::query()->findOrFail($id);
-        $subTask->delete();
-        return $subTask;
+        
+        if ($userId === $subTask->user_id) {
+            $subTask->delete();
+            return $subTask;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 }
